Complete change in approach
Instead of modifying the whole wdiget, only the building of the KNNIndex, part of the perplexity based
caclulation of nearest neighbors will be modified. This target change allows the forgoing of creating a
pip package since there is no need for compilation.

In process of the Orange update the whole project has been mistakenly deleted.
This allows for a restart and brings thus a change in folder structure.

The current task is to implement the ann-algorithm-api's in the KNNIndex build.